+++ CHANGELOG.md
@@ -3,7 +3,20 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- [#7](https://github.com/sunsided/async-tempfile-rs/pull/7):
+  Added `TempDir` for automatically deleted temporary directories.
+
+### Changed
+
+- Changed function signatures to take `Borrow<Path>` instead of `PathBuf`.
+
 ## [0.5.0] - 2023-12-06
+
+[0.5.0]: https://github.com/sunsided/async-tempfile-rs/releases/tag/0.5.0
 
 ### Changed
 
@@ -16,6 +29,8 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Some unnecessary heap allocations were removed.
 
 ## [0.4.0] - 2023-06-16
+
+[0.4.0]: https://github.com/sunsided/async-tempfile-rs/releases/tag/0.4.0
 
 ### Added
 
@@ -44,6 +59,3 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Internal
 
 - 🎉 Initial release.
-
-[0.5.0]: https://github.com/sunsided/async-tempfile-rs/releases/tag/0.5.0
-[0.4.0]: https://github.com/sunsided/async-tempfile-rs/releases/tag/0.4.0